refactor: update GuiElement class to add and remove CSS classes
This commit updates the GuiElement class in the src/gui.js file. It introduces two new methods, addClass and removeClass, which allow adding and removing CSS classes to the element. These methods take a class selector as an argument and update the element's classList accordingly. This change improves the flexibility and ease of use of the GuiElement class when working with CSS styles.