refactor: Add append method to ButtonGroup class
This commit adds a new method, `append`, to the `ButtonGroup` class. The `append` method allows for appending a single button or an array of buttons to the button group. It checks if the button is an instance of `BaseButton` and throws a `TypeError` if it is not. This method provides more flexibility in adding buttons to the button group.

Refactor the `ButtonGroup` class to include the `append` method for adding buttons.