refactor: Add markdownToHtml function and update Statusbar class
This commit adds the markdownToHtml function to the common.js file and updates the Statusbar class in gui.js. The markdownToHtml function is imported from common.js and used to convert markdown text to HTML. In the Statusbar class, the addStatus method is added to allow adding status elements with markdown content. This improves the functionality and flexibility of the Statusbar class.